Add an option for including the fade class.
The fade class triggers animations that have the modal slide down and
fade in when opened and slide up and fade out when closed. In order to
properly handle this, removal of the modal is now performed in an event
handler that listens for the "hidden" event fired by the modal. This
also fixes an issue with the modal not being completely removed when
the backdrop is clicked.